requirements.txt
streamlit
google-genai
pydantic
numpy
pandas
scikit-learn
pymupdf
python-dotenv
